dll_0000_gameui pauseMenuFn_80129ee0: recover u8 narrow on menu-select compare (99.15%->99.24%)
The (s8)lbl_803DBA64 != prev compare re-narrows prev to u8 in the target
(clrlwi). Cast the u8 local explicitly ((u8)prev) so MWCC emits the
zero-extend on both operands, matching the target byte compare.

+++ src/main/dll/dll_0000_gameui.c
@@ -2862,7 +2862,7 @@ void pauseMenuFn_80129ee0(void)
                     {
                         lbl_803DBA64 = menuMin;
                     }
-                    if ((s8)lbl_803DBA64 != prev)
+                    if ((s8)lbl_803DBA64 != (u8)prev)
                     {
                         Sfx_PlayFromObject(0, SFXTRIG_menu_fox_select);
                     }
